Move sessions-related queries to separate subfolder
