+++ posthog/queries/sessions/sessions.py
